# Cache HMAC object and fix memory leak