This PR improves `NewHMAC` so it reuses the same `EVP_MAC` object instead of fetching it every time.

Doing so also fixed a memory leak: the object returned by `EVP_MAC_fetch` should have been freed using `EVP_MAC_free`. The new approach doesn't require freeing that object, as it is kept as a global variable to be reused later on.

Thanks to caching, `NewHMAC` is now ~9% faster:

```go
goos: linux
goarch: amd64
pkg: github.com/golang-fips/openssl
cpu: AMD EPYC 7763 64-Core Processor                
           │   old.txt   │              new.txt               │
           │   sec/op    │   sec/op     vs base               │
NewHMAC-16   4.616µ ± 1%   4.199µ ± 1%  -9.04% (p=0.000 n=20)
```